+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Remove Deprecated Volume and ChangeDir methods. Fixes #236
 
 ## [v7.8.1] - 2025-08-07
 ### Fixed